requirements.txt
pandas>=2.2.3
tabulate>=0.9.0
matplotlib>=3.10.0
flask>=3.1.0
mysql-connector-python>=9.2.0